+++ src/core/factories/interactions.ts
@@ -1,6 +1,5 @@
 import { IUIInteractionProvider } from "../models/interfaces";
 import { CuiInteractionsType } from "../utlis/types";
-import { DefaultSetup } from "../../app/defaults/setup";
 import { FastDom, SyncInteractions } from "../utlis/interactions";
 
 export class CuiInteractionsFactory {
@@ -8,8 +7,8 @@ export class CuiInteractionsFactory {
      * Gets new instance of component focused logger
      * @param type - Interactions type
      */
-    public static get(type?: CuiInteractionsType): IUIInteractionProvider {
-        const interactionType = type ?? DefaultSetup.interaction;
+    public static get(type: CuiInteractionsType): IUIInteractionProvider {
+        const interactionType = type;
         switch (interactionType) {
             case 'async':
                 return new FastDom();