+++ src/core/factories/logger.ts
@@ -1,7 +1,7 @@
 import { CuiLogger } from "../utlis/logger";
 import { CuiLogLevel } from "../utlis/types";
 import { ICuiLogger } from "../models/interfaces";
-import { DefaultSetup } from "../../app/defaults/setup";
+import { STATICS } from "../utlis/statics";
 
 /**
  * 
@@ -12,6 +12,6 @@ export class CuiLoggerFactory {
      * @param name - component name
      */
     public static get(name: string, logLevel?: CuiLogLevel): ICuiLogger {
-        return new CuiLogger(name, logLevel ?? DefaultSetup.logLevel)
+        return new CuiLogger(name, logLevel ?? STATICS.logLevel)
     }
 }